REFACTOR: 그리드, 스레드를 컴포넌트로 전환
